Cria função de deletar posts, inclui data atual nas postagens e faz pequenas melhorias.
