+++ src/index.html
@@ -1,19 +1,22 @@
 <!DOCTYPE html>
 <html lang="pt-br">
+
     <head>
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <meta http-equiv="X-UA-Compatible" content="ie=edge" />
         <title>Vanellen🎞️</title>
-        <link rel="stylesheet" href="./login/login.css" />
-        <link rel="stylesheet" href="./register/register.css" />
-        <link rel="stylesheet" href="./initialPage/page.css" />
-        <link rel="stylesheet" href="./about/style.css" />
-        <link rel="stylesheet" href="./profile/style.css" />
+
+        <link rel="stylesheet" href="./pages/login/login.css" />
+        <link rel="stylesheet" href="./pages/register/register.css" />
+        <link rel="stylesheet" href="./pages/initialPage/page.css" />
+        <link rel="stylesheet" href="./pages/about/style.css" />
+        <link rel="stylesheet" href="./pages/profile/style.css" />
+
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     </head>
 
-    <body>
+    <body>  
         <main id="root"></main>
 
         <script type="module" src="./routes.js"></script>
@@ -24,5 +27,5 @@
         <script type="module" src="https://www.gstatic.com/firebasejs/9.9.3/firebase-firestore-compat.js"></script>
         <script src="https://www.gstatic.com/firebasejs/9.9.3/firebase-storage-compat.js"></script>
     </body>
-</html>
 
+</html>
